requirements.txt
annotated-types==0.6.0
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
defusedxml==0.7.1
Django==5.0.6
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==24.3
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-filter==24.2
django-js-asset==2.2.0
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-tinymce==4.0.0
django-treebeard==4.7.1
djangorestframework==3.15.1
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
idna==3.7
l18n==2021.3
laces==0.1.1
openpyxl==3.1.2
pillow==10.3.0
pillow_heif==0.16.0
pydantic==2.7.1
pydantic_core==2.18.2
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
telepath==0.3.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wagtail==6.1
wagtailcolumnblocks==2.3.2
Willow==1.8.0